Logic to perform arthimetic and clean the code
Calculator is able to take in the string that the users puts via
buttons. The string is sliced into three strings to be used in
the calculations.

The operate function was missing returns.

The isDigit function removed as it was not needed.

The operatorPressed function now checks if there is another operator
before it to avoid input multiple operatores.

Changed ' to " to remain consistant.